+++ modules/fimo_base/src/base/api.rs
@@ -1,0 +1,9 @@
+mod library;
+mod module;
+mod sys;
+mod version;
+
+pub use library::LibraryAPI;
+pub use module::ModuleAPI;
+pub use sys::{ExitStatus, SysAPI};
+pub use version::VersionAPI;
