+++ modules/fimo_base/src/base/api/version.rs
@@ -1,4 +1,4 @@
-use crate::base_api::{DataGuard, Locked, Unlocked};
+use crate::base::{DataGuard, Locked, Unlocked};
 use emf_core_base_rs::ffi::version::{ReleaseType, Version};
 use emf_core_base_rs::ownership::Owned;
 use emf_core_base_rs::Error;
